requirements.txt
tensorflow == 2.16.1
pyautogui == 0.9.54
pillow == 10.2.0
scikit-learn == 1.4.1.post1
pytesseract == 0.3.10
opencv-python == 4.9.0.80

